Use PackedCoordinateSequence for Geometry deserialization
When deserializing JTS Geometries, previously we made an
array of Coordinate objects.  Instead, use a PackedCoordinateSequence,
which is a double[] array under the hood. This reduces GC pressure,
total memory, and speeds up operations (probably due to CPU caching).

Do the same for serializing LineStrings and MultiLineStrings.
MultiPoints don't expose their coordinates as a CoordinateSequence.
Polygons/MultiPolygons need to be canonicalized, and converting
that requires more work.